Split examples into separate sections.
