+++ opentracing-examples/src/test/java/io/opentracing/examples/promise_propagation/Promise.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing.examples.async_propagation;
+package io.opentracing.examples.promise_propagation;
 
 import io.opentracing.ActiveSpan;
 import io.opentracing.mock.MockTracer;
