+++ opentracing-examples/src/test/java/io/opentracing/examples/suspend_resume_propagation/SuspendResume.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing.examples.async_propagation;
+package io.opentracing.examples.suspend_resume_propagation;
 
 import io.opentracing.ActiveSpan;
 import io.opentracing.mock.MockTracer;
