+++ opentracing-examples/src/test/java/io/opentracing/examples/suspend_resume_propagation/SuspendResumePropagationTest.java
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2016-2017 The OpenTracing Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package io.opentracing.examples.suspend_resume_propagation;
+
+import io.opentracing.mock.MockSpan;
+import io.opentracing.mock.MockTracer;
+import io.opentracing.mock.MockTracer.Propagator;
+import io.opentracing.util.ThreadLocalActiveSpanSource;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.List;
+import java.util.concurrent.Phaser;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * These tests are intended to simulate the kind of async models that are common in java async
+ * frameworks.
+ *
+ * For improved readability, ignore the phaser lines as those are there to ensure deterministic
+ * execution for the tests without sleeps.
+ *
+ * @author tylerbenson
+ */
+public class SuspendResumePropagationTest {
+
+  private final MockTracer tracer =
+      new MockTracer(new ThreadLocalActiveSpanSource(), Propagator.TEXT_MAP);
+  private Phaser phaser;
+
+  @Before
+  public void before() {
+    phaser = new Phaser();
+  }
+
+  @Test
+  public void testContinuationInterleaving() {
+    SuspendResume job1 = new SuspendResume(1, tracer);
+    SuspendResume job2 = new SuspendResume(2, tracer);
+
+    // Pretend that the framework is controlling actual execution here.
+    job1.doPart("some work for 1");
+    job2.doPart("some work for 2");
+    job1.doPart("other work for 1");
+    job2.doPart("other work for 2");
+    job2.doPart("more work for 2");
+    job1.doPart("more work for 1");
+
+    job1.done();
+    job2.done();
+
+    List<MockSpan> finished = tracer.finishedSpans();
+    assertThat(finished.size()).isEqualTo(2);
+
+    assertThat(finished.get(0).operationName()).isEqualTo("job 1");
+    assertThat(finished.get(1).operationName()).isEqualTo("job 2");
+
+    assertThat(finished.get(0).parentId()).isEqualTo(0);
+    assertThat(finished.get(1).parentId()).isEqualTo(0);
+  }
+}
